bcmdhd: Changed wakepkt_dump control flag to DHD_RPM_VAL
[Issue]
Hide the mac address printed as prhex in wakepkt_dump.

[Analysis]
We have DHD_RPM_VAL flag as logging control flag. This flag is supposed to control enable/disable PM logs in runtime.
and it is NOT enabled as default only Pixel platform. In other platform including hikey, PM logs can be seen since this flag is enabled.

[Solution & Fix]
1. dhd_msg_level should be uint32. not uint8.
   - 0x0001(DHD_ERROR_VAL) ~ 0x80000000(DHD_IOVAR_MEM_VAL)

2. added DHD_RPM_VAL condition to dhd_prhex()
   - DHD_RPM_VAL is used with DHD_RPM macro

3. Changed control flag from DHD_ERROR_VAL to DHD_RPM_VAL to control enable/disable "[wakepkt_dump]" log in Pixel platform.

Bug: 250885801
Test: Verified on Pixel device. prhex doesn't called as default

- Default (DHD_RPM_VAL is disabled)
[  110.031859] [03:11:43.545972][dhd][wlan]##### dhdpcie_host_wake caused by packets
[  110.031914] [03:11:43.546030][dhd][wlan]config check in_suspend: 1
[  110.031928] [03:11:43.546043][dhd][wlan]arp hmac_update:1
[  110.031945] [03:11:43.546060][dhd][wlan]wake 4 rxwake 4 readctrlwake 0
[  110.031959] [03:11:43.546074][dhd][wlan]unicast 4 muticast 0 broadcast 0 icmp 4
[  110.031972] [03:11:43.546087][dhd][wlan]multi4 0 multi6 0 icmp6 0 multiother 0
[  110.031984] [03:11:43.546100][dhd][wlan]icmp6_ra 0, icmp6_na 0, icmp6_ns 0
[  110.031997] [03:11:43.546112][dhd][wlan]Src_mac: 78:xx:xx:xx:xa:20, Dst_mac: c2:xx:xx:xx:xe:11

+++ dhd.h
@@ -4088,7 +4088,7 @@ void dhd_send_trap_to_fw_for_timeout(dhd_pub_t * pub, timeout_reasons_t reason);
 extern int dhd_bus_set_device_wake(struct dhd_bus *bus, bool val);
 extern void dhd_bus_dw_deassert(dhd_pub_t *dhd);
 #endif /* defined(PCIE_OOB) || defined(PCIE_INB_DW) */
-extern void dhd_prhex(const char *msg, volatile uchar *buf, uint nbytes, uint8 dbg_level);
+extern void dhd_prhex(const char *msg, volatile uchar *buf, uint nbytes, uint32 dbg_level);
 int dhd_tput_test(dhd_pub_t *dhd, tput_test_t *tput_data);
 void dhd_tput_test_rx(dhd_pub_t *dhd, void *pkt);
 #ifdef DHD_EFI
